recognition/arcface_torch/requirements.txt
easydict==1.9
numpy==1.19.5
requests==2.25.1
scikit-learn==0.24.2
torchsummary==1.5.1
mxnet